BuildContainer - Basic host-only container
Created a host only container with no networking configuration.

This doesn't run a container, it merely builds a container configuration sans networking. There's a bit of a TODO there with regards to finding the proper gateway configuration, a free port, and maybe some privacy concerns on the line.

Look forward to some follow on commits for github.com/docker/docker/api/types/network and NetworkingConfig

### Next steps
* Devine the gateway to use on the real host
* Set the HostConfig.NetworkMode to NetworkMode(bridge)
* Add nat.Portmap to HostConfig.PortBindings

DNS should be laid into the HostConfig as well