Readme: re code verbosity
Writing code is like a shower or a long drive/ride for me. The more time spent in it the more I reflect as I'm writing. When my tooling reminds me to make a decision about whether something should be public or, if so, comments are needed, the "Why?" of the experience isn't lost. 

The entirety of the code could be reduced to some variation of
```
handle(res, req) => {pullCode(req.url).then( bashExecThatShit)
```
Perhaps if this was "start-up" or "giving a talk" code, but this is gardening code. It grows wildly and organically, it understands it isn't on display at the MOMA nor powering air-traffic control. People have biases about how code should look, as do I. Code travels much more than people do and it should be able to tell its story to the people it travels to.